requirements.txt
bleach==1.5.0
Django==1.11.10
django-classy-tags==0.8.0
django-cms==3.4.5
django-cors-headers==2.4.0
django-filer==1.3.0
django-filter==2.0.0
django-formtools==2.1
django-mptt==0.8.7
django-polymorphic==1.3.1
django-sekizai==0.10.0
django-treebeard==4.1.2
djangocms-admin-style==1.2.7
djangocms-attributes-field==0.3.0
-e git://github.com/divio/djangocms-column.git@2f0a78cde7f5f34c1c925b649f8e2aa6b2ad8145#egg=djangocms_column
djangocms-file==2.0.2
djangocms-googlemap==1.1.1
djangocms-link==2.1.2
djangocms-picture==2.0.6
djangocms-snippet==1.9.2
djangocms-style==2.0.2
djangocms-text-ckeditor==3.5.0
djangocms-video==2.0.3
djangorestframework==3.8.2
earthengine-api==0.1.152
easy-thumbnails==2.4.2
google-api-python-client==1.7.7
html5lib==0.9999999
Markdown==2.6.11
numpy==1.15.1
oauth2client==4.1.2
olefile==0.45.1
pandas==0.23.4
Pillow==4.2.1
psycopg2==2.7.3.1
python-dateutil==2.7.3
pytz==2017.2
six==1.10.0
Unidecode==0.4.21
